% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_students.R
\name{can_source}
\alias{can_source}
\title{Check that .R file can be sourced without errors}
\usage{
can_source(files, ...)
}
\arguments{
\item{files}{a charachter vector of R file names to be sourced and checked if they can be run with no problem.}

\item{...}{not used.}
}
\value{
A data.frame with the name of the file, it's status (TRUE if was sourced properly, and FALSE otherwise),
and a note indicating possible issues.
}
\description{
The function gets a vector of .R file names and returns for each of them if it can be sourced or not.
This is helpful as an initial step before checking the homework (to make sure it can be loaded).
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
