% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hw.R
\name{check_hw}
\alias{check_hw}
\title{FUNCTION_TITLE}
\usage{
check_hw(hw_sub_dir = "", base_dir = getwd(),
  submissions_sub_dir = "submissions", sol_file, tests_to_run,
  create_grade_files = TRUE, unzip_submissions = TRUE,
  submission_file_ext_to_keep = c("R", "zip"), catch_copycats = TRUE,
  max_grade = 100, ...)
}
\arguments{
\item{hw_sub_dir}{PARAM_DESCRIPTION, Default: ''}

\item{base_dir}{PARAM_DESCRIPTION, Default: getwd()}

\item{submissions_sub_dir}{PARAM_DESCRIPTION, Default: 'submissions'}

\item{sol_file}{PARAM_DESCRIPTION}

\item{tests_to_run}{PARAM_DESCRIPTION}

\item{create_grade_files}{PARAM_DESCRIPTION, Default: TRUE}

\item{unzip_submissions}{PARAM_DESCRIPTION, Default: TRUE}

\item{submission_file_ext_to_keep}{PARAM_DESCRIPTION, Default: c("R", "zip")}

\item{catch_copycats}{PARAM_DESCRIPTION, Default: TRUE}

\item{max_grade}{PARAM_DESCRIPTION, Default: 100}

\item{...}{PARAM_DESCRIPTION}

\item{grades}{PARAM_DESCRIPTION}

\item{hw_sub_dir}{PARAM_DESCRIPTION}

\item{grades_sub_dir}{PARAM_DESCRIPTION, Default: 'grades'}

\item{char_to_keep}{PARAM_DESCRIPTION, Default: 5}

\item{get_id_from_file_name}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tools]{fileutils}}

\code{\link[tools]{fileutils}}
}
