% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{convert_to_eur}
\alias{convert_to_eur}
\title{Return EUR/USD value rate}
\usage{
convert_to_eur(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
numeric value
}
\description{
Return EUR/USD value rate
}
