% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copycats.R
\name{copycats_trap}
\alias{copycats_trap}
\alias{copycats_find}
\title{Help catch copycats}
\usage{
copycats_trap(file, ...)

copycats_find(file, show_print = TRUE, ...)
}
\arguments{
\item{file}{an .R file to update.}

\item{...}{not used}

\item{show_print}{logical (TRUE) if to print the rows with the suspected extra spaces.}
}
\value{
invisible TRUE. Also modifies the .R file that was in the input.
}
\description{
Takes a vector of .R files which includes solutions to homework
and adds 7 trailing spaces to each even line in the file.
This way, if the next semester you get homework which includes such a line,
it is clear that the student copied these homework from a solution another student
gave him from a previous year.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
