% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Generate Correlation Matrix For Given Year}
\usage{
corr_matrix(yr)
}
\arguments{
\item{yr}{The year of interest}
}
\value{
The correlation matrix of all numeric variables in data set for given year
}
\description{
Generate Correlation Matrix For Given Year
}
\examples{
corr_matrix(1995)
}
