% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hw.R, R/test_students.R
\name{create_grade_files}
\alias{create_grade_files}
\alias{create_grade_files_OLD}
\title{FUNCTION_TITLE}
\usage{
create_grade_files(grades, hw_sub_dir, grades_sub_dir = "grades",
  char_to_keep = 5, get_id_from_file_name = TRUE)

create_grade_files_OLD(results, HW_number, tests_to_run,
  grades_folder = "grades\\\\", char_to_trim = 6)
}
\arguments{
\item{grades}{PARAM_DESCRIPTION}

\item{hw_sub_dir}{PARAM_DESCRIPTION}

\item{grades_sub_dir}{PARAM_DESCRIPTION, Default: 'grades'}

\item{char_to_keep}{PARAM_DESCRIPTION, Default: 5}

\item{get_id_from_file_name}{PARAM_DESCRIPTION, Default: TRUE}

\item{results}{PARAM_DESCRIPTION}

\item{HW_number}{PARAM_DESCRIPTION}

\item{tests_to_run}{PARAM_DESCRIPTION}

\item{grades_folder}{PARAM_DESCRIPTION, Default: grades}

\item{char_to_trim}{PARAM_DESCRIPTION, Default: 6}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tools]{fileutils}}
}
