% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawname.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Plot the evolution of a name given in France}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{Name to look up}

\item{the_sex}{Sex of the name}
}
\value{
a graphic
}
\description{
Plot the evolution of a name given in France
}
\examples{
\dontrun{
draw_a_name("Hamza","M")
}

}
