% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Plots several name distribution over the years}
\usage{
draw_names(vector_name)
}
\arguments{
\item{vector_name}{Vector composed of names}
}
\value{
a graph
}
\description{
Plots several name distribution over the years
}
