% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{eur}
\alias{eur}
\title{Return string with euro mark from numeric value}
\usage{
eur(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
character string
}
\description{
Return string with euro mark from numeric value
}
