% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_students.R
\name{file_ext_to_keep}
\alias{file_ext_to_keep}
\alias{only_R_files}
\title{Get only .R files}
\usage{
file_ext_to_keep(files, file_ext = c("R"), case_sensitive = FALSE)

only_R_files(files, case_sensitive = FALSE)
}
\arguments{
\item{files}{- a charachter vector of file names}

\item{case_sensitive}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
only files which are R/r files.
}
\description{
Give a vector of possible file names, returns only the ones that are .R files.
}
\examples{

files <- c("a", "b.R", "c.RR", "d.Rdata", "e.R")
only_R_files(files)

}
\seealso{
\code{\link[tools]{file_ext}}
}
