% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework.R
\docType{package}
\name{Homework}
\alias{Homework}
\alias{Homework-package}
\title{Stores 5 functions for database manipulation}
\description{
Enable to load, save and modify excel, csv files, and to work on the French first name database
}
\author{
Raphael Selignan <raphael.selignan@polytechnique.edu>
}
