% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrawnames.R
\name{mdrawnames}
\alias{mdrawnames}
\title{Draw a plot comparing the evolutions of multiple names in France}
\usage{
mdrawnames(vector_name)
}
\arguments{
\item{vector_name}{vector of the names to be plotted}
}
\value{
a plot
}
\description{
Draw a plot comparing the evolutions of multiple names in France
}
\examples{
\dontrun{
mdrawnames(c("Hamza", "Alex", "Raphael"))
}
}
