% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjs_pts.R
\name{mjs_pts}
\alias{mjs_pts}
\title{Michael Jordan Points}
\usage{
mjs_pts(yr)
}
\arguments{
\item{yr}{The year of interest}
}
\value{
The number of points Michael Jordan scored in \code{yr} year
}
\description{
Michael Jordan Points
}
\examples{
mjs_pts(1995)
}
