% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{import all csv files from a given folder}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{directory of the folder}
}
\value{
list
}
\description{
import all csv files from a given folder
}
\examples{
\dontrun{
my_csv_reader("Users:/Documents")
}

}
