% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycsvreader.R
\name{mycsvreader}
\alias{mycsvreader}
\title{Import the csv files from a given folder}
\usage{
mycsvreader(folder)
}
\arguments{
\item{folder}{path of the folder to import}
}
\value{
a list
}
\description{
Import the csv files from a given folder
}
\examples{
\dontrun{
my_csv_reader(".../path")
}

}
