% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{neur}
\alias{neur}
\title{Return numeric value from euro value srting with euro mark.}
\usage{
neur(eurstring)
}
\arguments{
\item{eurstring}{character string}
}
\value{
character string
}
\description{
Return numeric value from euro value srting with euro mark.
}
