\name{pythag}
\alias{pythag}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Pythagorean Theorem Function
}
\description{This function solves for any side of a right triangle given the input of the other 2 sides, uses pythagorean theorem.
}
\usage{
pythag(a, b, c)
}

\arguments{
  \item{a}{ integer value of side 'a' or 'NULL'}
\item{b}{ integer value of side 'b' or 'NULL'}
\item{c}{ integer value of side 'c' or 'NULL'}
}

\value{
the value returned will be the integer value of the 'Null' side of the triangle.
}
\author{Cristin Kaspar
%%  ~~who you are~~
}
