% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readexcelmulti.R
\name{read_excel_multi}
\alias{read_excel_multi}
\title{Import all sheets from an excel file (xlsx)}
\usage{
read_excel_multi(file)
}
\arguments{
\item{file}{Path to the excel file to be loade}
}
\value{
a list
}
\description{
Import all sheets from an excel file (xlsx)
}
\examples{
\dontrun{
read_excel_multi(".../path/file.xlsx")
}

}
