% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rnormC}
\alias{rnormC}
\title{A normally distributed sampler using Rcpp}
\usage{
rnormC(E, se)
}
\arguments{
\item{E}{the mean of a normal distribution}

\item{se}{the standard error of a normal distribution}
}
\value{
a random sample of size 1
}
\description{
Generate a normally distributed random sample using Rcpp
}
\examples{
\dontrun{
rnC <- replicate(100,rnormC(5,1))
plot(rnC,type='l')
}
}
