% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runifC}
\alias{runifC}
\title{Random sampler between 0 and 1 using Rcpp}
\usage{
runifC(n)
}
\arguments{
\item{n}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
Generate n random numbers between 0 and 1 using Rcpp
}
\examples{
\dontrun{
rnC <- runifC(100)
plot(rnC,type='l')
}
}
