% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_MetropolisC}
\alias{rw_MetropolisC}
\title{A random walk Metropolis sampler using Rcpp}
\usage{
rw_MetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{given standard deviation}

\item{x0}{the initial value given}

\item{N}{the length of the resulting chain}
}
\value{
a list of a chain of size \code{n} and the number of times it is accepted during the process of using this method
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
N <- 2000
rwk <- rw_MetropolisC(2,20,N)
plot(rwk$x,type='l')
accept.rate <- rwk$k/N
}
}
