% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_MetropolisR.R
\name{rwM.gen.chain}
\alias{rwM.gen.chain}
\title{A random walk Metropolis sampler using R}
\usage{
rwM.gen.chain(sigma, x0, N)
}
\arguments{
\item{sigma}{given standard deviation}

\item{x0}{the initial value given}

\item{N}{the length of the resulting chain}
}
\value{
a list of a chain of size \code{n} and the accepte rate during the process of using this method
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using R
}
\examples{
\dontrun{
N <- 2000
rwk <- rw_MetropolisR(2,20,N)
plot(rwk$x,type='l')
accept.rate <- rwk$k/N
}
}
