% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savecsv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save Csv Files}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{Data to be saved as csv}

\item{filename}{Name chosen by the user for csv file}

\item{row.names}{Parameter to omit the row names}

\item{...}{Other parameters}
}
\value{
a csv file
}
\description{
Save Csv Files
}
\examples{
\dontrun{
save_as_csv(dataname, my_csv_file_name, row.names=FALSE)
}

}
