\name{StatComp20015-package}
\alias{StatComp20015-package}
\alias{StatComp20015}
\docType{package}
\title{
  Final project of Statistical Computation
}
\description{
  The R package contains several functions related to the SI model,SIS model, and SIR model.It also contains all of homeworks.
}
\author{
XuZiling, email optional.

Maintainer: XuZiling <xuziling@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}

