% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{substrRight}
\alias{substrRight}
\title{FUNCTION_TITLE}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
