% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{test_students}
\alias{test_students}
\title{FUNCTION_TITLE}
\usage{
test_students(hw_submitters, sol_file, tests_to_run, student_id_fun = NULL,
  timeout = 0.5, use_do.call, check_sol_fun = function(student_sol,
  teacher_sol) {     isTRUE(all.equal(student_sol, teacher_sol, tolerance =
  0.01, check.attributes = FALSE)) })
}
\arguments{
\item{hw_submitters}{PARAM_DESCRIPTION}

\item{sol_file}{PARAM_DESCRIPTION}

\item{tests_to_run}{PARAM_DESCRIPTION}

\item{student_id_fun}{PARAM_DESCRIPTION, Default: NULL}

\item{timeout}{PARAM_DESCRIPTION, Default: 0.5}

\item{use_do.call}{PARAM_DESCRIPTION}

\item{check_sol_fun}{PARAM_DESCRIPTION, Default:
function(student_sol, teacher_sol) {
  isTRUE(all.equal(
    student_sol, teacher_sol, tolerance = 0.01,
    check.attributes = FALSE
  ))
}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
