% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{usd_rate}
\alias{usd_rate}
\title{Get USD/Value rate}
\usage{
usd_rate(x, currency = "EUR")
}
\arguments{
\item{x}{numeric value}

\item{currency}{INFO}
}
\description{
Return info about curency value in USD. Deafult value is EUR.
}
