% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_oldest.R
\name{years_oldest}
\alias{years_oldest}
\title{Oldest Player That Year}
\usage{
years_oldest(yr)
}
\arguments{
\item{yr}{The year of interest}
}
\value{
The oldest player in the league in given year sorted alphabetically
}
\description{
Oldest Player That Year
}
\examples{
years_oldest(1995)
}
