% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_students.R
\name{create_grade_files}
\alias{create_grade_files}
\title{FUNCTION_TITLE}
\usage{
create_grade_files(results, HW_number, tests_to_run,
  grades_folder = "grades\\\\", char_to_trim = 6)
}
\arguments{
\item{results}{PARAM_DESCRIPTION}

\item{HW_number}{PARAM_DESCRIPTION}

\item{tests_to_run}{PARAM_DESCRIPTION}

\item{grades_folder}{PARAM_DESCRIPTION, Default: grades}

\item{char_to_trim}{PARAM_DESCRIPTION, Default: 6}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
