% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Plots the occurences of a name over the years}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{Name wanted}

\item{the_sex}{M or F for male or female}
}
\value{
a graph
}
\description{
Plots the occurences of a name over the years
}
\examples{
\dontrun{
draw_a_name("Marcel","M")
}
}
