% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamza.R
\docType{package}
\name{hamza-package}
\alias{hamza-package}
\alias{hamza}
\title{Toolkit functions to start with R - Hamza}
\description{
This package is a toolkit to help get started with R
}
\author{
Hamza <hamza.sentissi@polytechnique.edu>
}
