% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_multi.R
\name{read_excel_multi}
\alias{read_excel_multi}
\title{import all sheets from an excel file}
\usage{
read_excel_multi(file)
}
\arguments{
\item{file}{Name of the file from which extracting is needed}
}
\value{
a list
}
\description{
import all sheets from an excel file
}
\examples{
\dontrun{
read_excel_multi(file/path)
}

}
